feat(agent): runtime translation + per-agent scheduler (U6, U7)
Runtime ties the Pi supervisor, a persistence Store seam, and the WS broadcaster
into the Super Threads engine. Per-(session,agent) serial queue with the
scheduler as single owner of all terminal transitions + promotion (KTD12),
idempotent terminal (Pi completion vs process-exit double-fire), in-tx seq
allocation persist-first-then-broadcast (KTD6), idempotent action append, and
force-resolve of open actions on terminal. Process exits fail the running task
and promote the next. DBStore implements the seam over pgx; fake-Store +
real-supervisor tests cover enqueue/run/complete, queue+promote, concurrent
agents, exit-fails-running, and idempotent terminal. (cast AllocateEventSeq to
bigint.)

+++ server/internal/db/queries/tasks.sql
@@ -6,7 +6,7 @@ INSERT INTO session_event_seq (session_id, next_seq)
 VALUES ($1, 2)
 ON CONFLICT (session_id)
 DO UPDATE SET next_seq = session_event_seq.next_seq + 1
-RETURNING next_seq - 1 AS seq;
+RETURNING (next_seq - 1)::bigint AS seq;
 
 -- name: PeekEventSeq :one
 -- Highest allocated seq for a session (next_seq - 1), or 0 if none. Used as the
